Move MapExt to persistence module
