+++ src/persistent/mod.rs
@@ -8,10 +8,17 @@ use bevy::prelude::*;
 use serde_json::Value as JsonValue;
 
 mod io;
+mod map_ext;
+
+pub use map_ext::MapExt;
 
 pub fn plugin(app: &mut App) {
 	app.insert_resource(StoragePath::new());
 }
+
+/// System set where all persistent resources are fetched
+#[derive(SystemSet, Clone, Copy, PartialEq, Eq, Debug, Default, Hash)]
+pub struct LoadPersistentResourcesSystems;
 
 /// Extension trait that allows registering of resources that are bound to persistent state
 pub trait RegisterSaveableResource {
@@ -24,7 +31,10 @@ impl RegisterSaveableResource for App {
 	fn register_saveable_resource<T: Saveable + Resource>(&mut self) -> &mut Self {
 		self.init_resource::<T>()
 			.init_resource::<JsonStore<T>>()
-			.add_systems(Startup, load_system::<T>)
+			.add_systems(
+				Startup,
+				load_system::<T>.in_set(LoadPersistentResourcesSystems),
+			)
 			.add_systems(Update, save_system::<T>.run_if(resource_changed::<T>))
 			.add_systems(Update, write_system::<T>)
 	}
